requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
pydantic>=2.0.0
numpy>=1.21.0
pytest>=7.0.0 